clean up vocab usage
move vocab-related stuff into vocab.ts and move away from Ontola (ie, @ontologies) libraries

there's a subtle incompatibility between the `@ontologies`
version of `NamedNode` and the "standard" (ie, rdfjs-base https://github.com/rdfjs-base)
versions - the `@ontologies` version doesn't require an `equals` method, so trying to pass
`NamedNodes` from that library to the Inrupt libraries doesn't work.
Trying to switch would be a big lift for not much work and potentially
make us less compatible with `solid-client-js` so I opted to use
`@rdfjs/namespace`.

Unfortunately the most recent versions of the `@rdfjs` libraries have
been converted to ESM modules and are not transpiled to an older
version of JS like most NPM libraries, which makes Jest choke at the
moment, so I'm following the lead of `solid-client-js` and requiring an
older version of this lib.

The RDF ecosystem is a smoking wasteland of incompatibility and pain!
A real shame because RDF is awesome!