overhaul spaces setup function
using hooks makes the timing on all of this very weird, and the current implementation failed because various resources weren't set up yet. move to a mostly non-hook-based setup implementation.

I added a bunch of functions to hooks.ts that should be moved later, but leaving them here as I'm not done overhauling setup functions